+++ game-dragonquest7reimagined/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None
 
+## [0.1.1] - 2026-03-27
+
+- Fixed: Config and Save paths
+
 ## [0.1.0] - 2026-02-05
 
 - Initial release