+++ game-unrealtournament2004/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None Planned
 
+## [0.1.1] - 2026-03-27
+
+- Added: Button to browse and download OldUnreal patches from GitHub
+
 ## [0.1.0] - 2026-02-20
 
 - Inital Release